added nested serialization to answer

+++ surveys/api_surveys/serializers.py
@@ -1,12 +1,13 @@
 from rest_framework import serializers
 
-from api_surveys.models import *
+from api_surveys.models import Choices, Question, Survey, Answers, UserSurveyQuestion, UserSurvey
 
 
+# create survey
 class ChoicesCreateSerializers(serializers.ModelSerializer):
     class Meta:
         model = Choices
-        fields = ('id','choices_answer',)
+        fields = ('id', 'choices_answer',)
 
 
 class QuestionCreateSerializers(serializers.ModelSerializer):
@@ -14,7 +15,7 @@ class QuestionCreateSerializers(serializers.ModelSerializer):
 
     class Meta:
         model = Question
-        fields = ('id','question_text', 'question_type', 'choices')
+        fields = ('id', 'question_text', 'question_type', 'choices')
 
 
 class SurveyCreateSerializers(serializers.ModelSerializer):
@@ -22,7 +23,7 @@ class SurveyCreateSerializers(serializers.ModelSerializer):
 
     class Meta:
         model = Survey
-        fields = ('id','title', 'start_date', 'end_data', 'description', 'question')
+        fields = ('id', 'title', 'start_date', 'end_date', 'description', 'question')
 
     def create(self, validated_data):
         question_data = validated_data.pop('question')
@@ -38,10 +39,11 @@ class SurveyCreateSerializers(serializers.ModelSerializer):
         return id_survey
 
 
+# update survey
 class ChoicesUpdateDeleteSerializers(serializers.ModelSerializer):
     class Meta:
         model = Choices
-        fields = ('id','choices_answer',)
+        fields = ('id', 'choices_answer',)
 
 
 class QuestionUpdateDeleteSerializers(serializers.ModelSerializer):
@@ -49,7 +51,7 @@ class QuestionUpdateDeleteSerializers(serializers.ModelSerializer):
 
     class Meta:
         model = Question
-        fields = ('id','question_text', 'question_type', 'choices')
+        fields = ('id', 'question_text', 'question_type', 'choices')
 
 
 class SurveyUpdateDeleteSerializers(serializers.ModelSerializer):
@@ -57,20 +59,19 @@ class SurveyUpdateDeleteSerializers(serializers.ModelSerializer):
 
     class Meta:
         model = Survey
-        fields = ('id','title', 'start_date', 'end_data', 'description', 'question')
+        fields = ('id', 'title', 'start_date', 'end_date', 'description', 'question')
 
     def update(self, instance, validated_data):
         instance.title = validated_data.get('title', instance.title)
-        instance.start_date = validated_data.get('start_date', instance.start_date)
-        instance.end_data = validated_data.get('end_data', instance.end_data)
+        instance.end_date = validated_data.get('end_data', instance.end_date)
         instance.description = validated_data.get('description', instance.description)
-        print(instance.question.all())
+        if 'start_date' in validated_data:
+            raise serializers.ValidationError({'start_date': 'Error,start_date field cannot be changed'})
         for q in instance.question.all():
             if q.choices:
                 for c in q.choices.all():
                     c.delete()
             q.delete()
-        print(instance.question.all())
 
         for question in validated_data['question']:
             question_dict = dict(question)
@@ -81,31 +82,60 @@ class SurveyUpdateDeleteSerializers(serializers.ModelSerializer):
                 for choices in question['choices']:
                     Choices.objects.create(question_id=question_id, **dict(choices))
             instance.question.add(question_id)
-
-
-
-
-
-
-
-
-
-
-
-
-
-
         instance.save()
         return instance
 
 
+# survey active
 class SurveyViewActiveSerializers(serializers.ModelSerializer):
     class Meta:
         model = Survey
         fields = '__all__'
 
 
+# answer survey
+
 class AnswerCreateSerializers(serializers.ModelSerializer):
+    class Meta:
+        model = Answers
+        fields = ('answer_text', 'choices_answer_id')
+
+
+class UserSurveyQuestionCreateSerializers(serializers.ModelSerializer):
+    answers = AnswerCreateSerializers(many=True, required=False)
+
+    class Meta:
+        model = UserSurveyQuestion
+        fields = ('question_id', 'answers')
+
+
+class UserSurveyCreateSerializers(serializers.ModelSerializer):
+    user_survey_question = UserSurveyQuestionCreateSerializers(many=True, required=False)
+
+    class Meta:
+        model = UserSurvey
+        fields = ('user_id', 'survey_id', 'user_survey_question')
+
+    def create(self, validated_data):
+        user_survey_question_data = validated_data.pop('user_survey_question')
+        id_user_survey = UserSurvey.objects.create(**validated_data)
+
+        for user_survey_question in user_survey_question_data:
+            user_survey_question_dict = dict(user_survey_question)
+            if 'answers' in user_survey_question_dict:
+                del user_survey_question_dict['answers']
+
+            user_survey_question_id = UserSurveyQuestion.objects.create(user_survey_id=id_user_survey,
+                                                                        **user_survey_question_dict)
+            if 'answers' in dict(user_survey_question):
+                for user_survey in user_survey_question['answers']:
+                    Answers.objects.create(user_survey_question_id=user_survey_question_id, **dict(user_survey))
+        return id_user_survey
+
+
+# completed user surveys
+
+class CompletedUserSurveysSerializers(serializers.ModelSerializer):
     class Meta:
         model = Answers
         fields = '__all__'
